requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
